+++ apps/docs/src/components/Inputs/slider/MainView.tsx
@@ -1,22 +1,19 @@
-import { useCallback, useState } from "@elf-framework/sapa";
-import { Slider, VBox } from "@elf-framework/ui";
+import React from "react";
+import { Slider, VBox } from "@react-elf/ui";
 
 export function MainView() {
-  const [value, setValue] = useState(0);
-  const onInput = useCallback(
-    (v) => {
-      setValue(v);
-    },
-    [setValue]
-  );
   return (
     <VBox
       style={{
         gap: 30,
         backgroundColor: "var(--color-gray-0)",
         padding: [60, 200],
-        position: "relative",
         alignItems: "center",
+        fontSize: 13,
+        position: "relative",
+        margin: [0, "auto"],
+        justifyContent: "center",
+        height: 500,
       }}
     >
       <div
@@ -25,27 +22,27 @@ export function MainView() {
           flexDirection: "column",
           gap: 30,
           fontSize: 20,
-          width: "100%",
+          width: "80%",
         }}
       >
         <Slider
           min={0}
           max={10000}
           step={20}
-          value={value}
-          onInput={onInput}
+          defaultValue={0}
           valuePlacement="top"
           label={"Contrast"}
           // eslint-disable-next-line no-undef
-          valueFunc={(v) => new Intl.NumberFormat().format(v)}
+          valueFunc={(v) => {
+            return new Intl.NumberFormat().format(v);
+          }}
         />
 
         <Slider
           min={0}
           max={10000}
           step={20}
-          value={value}
-          onInput={onInput}
+          defaultValue={0}
           valuePlacement="top"
           showValue
           // eslint-disable-next-line no-undef
@@ -56,8 +53,7 @@ export function MainView() {
           min={0}
           max={10000}
           step={20}
-          value={value}
-          onInput={onInput}
+          defaultValue={0}
           showValue
           // eslint-disable-next-line no-undef
           valueFunc={(v) => new Intl.NumberFormat().format(v)}
@@ -67,8 +63,7 @@ export function MainView() {
           min={0}
           max={10000}
           step={20}
-          value={value}
-          onInput={onInput}
+          defaultValue={0}
           valuePlacement="bottom"
           showValue
           // eslint-disable-next-line no-undef