+++ apps/docs/src/components/Inputs/slider/ValueFormatView.tsx
@@ -1,4 +1,5 @@
-import { VBox, Slider } from "@elf-framework/ui";
+import React from "react";
+import { VBox, Slider } from "@react-elf/ui";
 
 export function ValueFormatView() {
   return (
@@ -20,7 +21,7 @@ export function ValueFormatView() {
       >
         <Slider
           label={"Opacity"}
-          value={40}
+          defaultValue={40}
           min={0}
           max={100}
           step={1}
@@ -29,7 +30,7 @@ export function ValueFormatView() {
 
         <Slider
           label={"Exposure"}
-          value={1.83}
+          defaultValue={1.83}
           fill
           min={-5}
           max={5}
@@ -39,7 +40,7 @@ export function ValueFormatView() {
 
         <Slider
           label={"Expense"}
-          value={20}
+          defaultValue={20}
           fill
           min={0}
           max={100}