+++ packages/@react-elf/ui/src/index.ts
@@ -21,3 +21,4 @@ export * from "@react-elf/radio";
 export * from "@react-elf/help-text";
 export * from "@react-elf/input-editor";
 export * from "@react-elf/switch";
+export * from "@react-elf/slider";